+++ EventShadow.API/Models/EventDevices.cs
@@ -7,7 +7,7 @@ namespace EventShadow.API.Models
     {
         public long Id { get; set; }
         public long EventId { get; set; }
-        public int EventShadowId { get; set; }
+        public long EventShadowId { get; set; }
 
         public Events Event { get; set; }
         public EventShadowDevices EventShadow { get; set; }
